Global Super Store Dataset
https://www.kaggle.com/datasets/apoorvaappz/global-super-store-dataset?select=Global_Superstore2.csv

Context
Shopping online is currently the need of the hour. Because of this COVID, it's not easy to walk in a store randomly and buy anything you want. I this I am trying to understand a few things like

Customers Analysis
Profile the customers based on their frequency of purchase - calculate frequency of purchase for each customer
Do the high frequent customers are contributing more revenue
Are they also profitable - what is the profit margin across the buckets
Which customer segment is most profitable in each year.
How the customers are distributed across the countries- -
Product Analysis
Which country has top sales?
Which are the top 5 profit-making product types on a yearly basis
How is the product price varying with sales - Is there any increase in sales with the decrease in price at a day level
What is the average delivery time across the counties - bar plot
I will keep updating the analysis.

Content
Once you download the file the rows you see are the details of the order done online by people across the globe in the time frame 1-jan-2011 to 31-dec-2014. There are no missing values in the majority of columns except postal code, you can drop it if not required.

Acknowledgements
I am thankfull for all the resources I came across on this dataset on Kaggle. I have learnt a lot from all of them.

Inspiration
If you are a beginner in data science and if you want to know how e-commerce analytics work then this is a great place to start understanding it. You can get many insights from this dataset. I am looking forward to seeing your analytical approach on this dataset. All the best :)